Вынес тесты на ресурсы в отдельный файл test_get_resource_data.py
